* win32 warning ** fixed warning C4389: '!=' : signed/unsigned mismatch (6) ** fixed warning C4307: integral constant overflow warning disable (2) ** fixed warning C4242: '=' : conversion from 'sector_t' to 'long', possible loss of data (74) ** fixed warning C4146: unary minus operator applied to unsigned type, result still unsigned (1) ** fixed warning C4133: '=' : incompatible types - from - to - (1) ** fixed warning C4018: signed/unsigned mismatch (5)

+++ drbd/drbd_nl.c
@@ -648,7 +648,7 @@ static char **make_envp(struct env *env)
 	env->buffer[env->pos++] = 0;
 	for (b = env->buffer, n = 1; *b; n++)
 		b = strchr(b, 0) + 1;
-	if (env->size - env->pos < sizeof(envp) * n)
+	if (env->size - env->pos < (int)(sizeof(envp) * n))
 		goto out_nomem;
 	envp = (char **)(env->buffer + env->size) - n;
